+++ Framework/Utils/src/esd-2-arrow-converter.cxx
@@ -50,9 +50,9 @@ WorkflowSpec defineDataProcessing(ConfigContext const& specs)
     readers::AODReaderHelpers::run2ESDConverterCallback(),
     {ConfigParamSpec{"esd-file", VariantType::String, "AliESDs.root", {"Input ESD file"}},
      ConfigParamSpec{"events", VariantType::Int, 0, {"Number of events to process (0 = all)"}},
-     ConfigParamSpec{"start-value-enumeration", VariantType::Int, 0, {"initial value for the enumeration"}},
-     ConfigParamSpec{"end-value-enumeration", VariantType::Int, -1, {"final value for the enumeration"}},
-     ConfigParamSpec{"step-value-enumeration", VariantType::Int, 1, {"step between one value and the other"}}}};
+     ConfigParamSpec{"start-value-enumeration", VariantType::Int64, 0ll, {"initial value for the enumeration"}},
+     ConfigParamSpec{"end-value-enumeration", VariantType::Int64, -1ll, {"final value for the enumeration"}},
+     ConfigParamSpec{"step-value-enumeration", VariantType::Int64, 1ll, {"step between one value and the other"}}}};
 
   return {run2Converter};
 }
